refactor(web): remove nested /components folder in Dropdown
